refactoring after review 2
