+++ server/src/streams/FileTargetsWriter.cpp
@@ -1,10 +1,10 @@
 #include "FileTargetsWriter.h"
 
 void FileTargetsWriter::writeResponse(
-    const std::vector<std::shared_ptr<BuildDatabase::TargetInfo>> &targets,
+    const std::vector<fs::path> &targetPaths,
     const utbot::ProjectContext &projectContext) {
     if (!hasStream()) {
         return;
     }
-    writeTargets(targets, projectContext);
+    writeTargets(targetPaths, projectContext);
 }
